Cleaning debug code
Former-commit-id: 2a822d1df5eda2223a0499a3e146857a4202113b

+++ source/mps_routines/mps_tools/finite/debug.cpp
@@ -45,10 +45,10 @@ void MPS_Tools::Finite::Debug::check_integrity_of_sim(const class_finite_chain_s
     if(state.get_length() != superblock.get_length())
         throw std::runtime_error("Length mismatch in state and superblock: " + std::to_string(state.get_length()) + " " + std::to_string(superblock.get_length()));
 
-    if(state.get_length()  != superblock.environment_size + 2 )
+    if(state.get_length()  !=  superblock.environment_size + 2 )
         throw std::runtime_error("Length mismatch in state and superblock env + 2: " + std::to_string(state.get_length()) + " " + std::to_string(superblock.environment_size + 2 ));
 
-    if(state.get_position()  != superblock.Lblock->size )
+    if(state.get_position()  != (int) superblock.Lblock->size )
         throw std::runtime_error("Mismatch in state position and sites in superblock env: " + std::to_string(state.get_position()) + " " + std::to_string(superblock.Lblock->size));
 
 
@@ -83,10 +83,10 @@ void MPS_Tools::Finite::Debug::check_integrity_of_mps(const class_finite_chain_s
         if(state.get_ENV2_L().size() + state.get_ENV2_R().size() != state.get_length())
             throw std::runtime_error("Mismatch in ENV size: " + std::to_string(state.get_ENV_L().size() + state.get_ENV_R().size()) + " " + std::to_string(state.get_length()));
 
-        if(state.get_ENV_L().back().size != state.get_position())
+        if((int) state.get_ENV_L().back().size != state.get_position())
             throw std::runtime_error("Mismatch in ENV_L size and position: " + std::to_string(state.get_ENV_L().back().size) + " " + std::to_string(state.get_position()));
 
-        if(state.get_ENV_R().front().size != state.get_length() - state.get_position() - 2)
+        if((int) state.get_ENV_R().front().size != (int)(state.get_length() - state.get_position() - 2))
             throw std::runtime_error("Mismatch in ENV_R size+1 and length-position: " + std::to_string(state.get_ENV_R().front().size) + " " + std::to_string(state.get_length() - state.get_position()-2));
 
         spdlog::debug("\tChecking matrix sizes on the left side");
